backend/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
flake8==4.0.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
zope.event==4.5.0
zope.interface==5.4.0
